fix: handle serve flags through command
